+++ plugins/newspack-plugin/src/wizards/audience/views/content-gates/advanced-settings.tsx
@@ -8,6 +8,7 @@
 import { __ } from '@wordpress/i18n';
 import { ToggleControl, __experimentalHStack as HStack, __experimentalVStack as VStack } from '@wordpress/components'; // eslint-disable-line @wordpress/no-unsafe-wp-apis
 import { useDispatch } from '@wordpress/data';
+import { decodeEntities } from '@wordpress/html-entities';
 import { useEffect, useRef, useState } from '@wordpress/element';
 
 /**
@@ -24,7 +25,7 @@ const AdvancedSettings = ( { closeModal, showModal }: { closeModal: () => void; 
 	const initialConfig = {
 		...( wizardData?.config?.advanced_settings || {} ),
 	};
-	const { wizardApiFetch, isFetching, resetError, setError } = useWizardApiFetch( AUDIENCE_CONTENT_GATES_WIZARD_SLUG );
+	const { wizardApiFetch, isFetching, resetError } = useWizardApiFetch( AUDIENCE_CONTENT_GATES_WIZARD_SLUG );
 	const { addNotice, resetNotices, updateWizardSettings } = useDispatch( WIZARD_STORE_NAMESPACE );
 	const [ config, setConfig ] = useState< AdvancedSettingsConfig >( initialConfig );
 
@@ -70,7 +71,11 @@ const AdvancedSettings = ( { closeModal, showModal }: { closeModal: () => void; 
 					} );
 				},
 				onError: ( fetchError: WpFetchError ) => {
-					setError( fetchError );
+					addNotice( {
+						message: decodeEntities( fetchError.message ),
+						type: 'error',
+						id: 'content-gates-advanced-settings-error',
+					} );
 				},
 				onFinally: () => {
 					closeModal();