+++ plugins/newspack-plugin/CHANGELOG.md
@@ -1,3 +1,10 @@
+## newspack [6.45.1](https://github.com/Automattic/newspack-workspace/compare/newspack@6.45.0...newspack@6.45.1) (2026-07-09)
+
+
+### Bug Fixes
+
+* **wizards:** stop wizard error notices from flickering (NPPM-2733) ([#513](https://github.com/Automattic/newspack-workspace/issues/513)) ([3600da0](https://github.com/Automattic/newspack-workspace/commit/3600da05fe06fc37446e409dbea13eb245e851cc))
+
 # newspack [6.45.0](https://github.com/Automattic/newspack-workspace/compare/newspack@6.44.9...newspack@6.45.0) (2026-07-08)